+++ scriptcwl/step.py
@@ -1,6 +1,6 @@
 import os
 import six
-from urlparse import urlparse
+from six.moves.urllib.parse import urlparse
 from ruamel.yaml.comments import CommentedMap
 
 from cwltool.load_tool import fetch_document, validate_document
@@ -87,6 +87,9 @@ class Step(object):
             raise ValueError('Invalid input "{}"'.format(name))
         self.step_inputs[name] = value
 
+    def _set_name_in_workflow(self, name):
+        self.name_in_workflow = name
+
     def output_to_input(self, name):
         """Convert the name of an output to an input for a next Step.
 
@@ -101,7 +104,7 @@ class Step(object):
         """
         if name not in self.output_names:
             raise ValueError('Invalid output "{}"'.format(name))
-        return ''.join([self.name, '/', name])
+        return ''.join([self.name_in_workflow, '/', name])
 
     def _input_optional(self, inp):
         """Returns True if a step input parameter is optional.
@@ -163,7 +166,7 @@ class Step(object):
             str containing all input names and types.
         """
         doc = []
-        for inp, typ in self.input_types.iteritems():
+        for inp, typ in self.input_types.items():
             if isinstance(typ, six.string_types):
                 typ = "'{}'".format(typ)
             doc.append('{}: {}'.format(inp, typ))