+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 0.5.1
+
+### Added
+
+* Allow addition of default values to workflow inputs (#32)
+* List of steps and workflows in steps library is ordered alphabetically
+
+## 0.5.0
+
+### Added
+
+* Python 3 compatibility (and testing with [tox](https://tox.readthedocs.io/en/latest/))
+
+## 0.4.0
+
+### Added
+
+* Generate unique names for steps that are added to the workflow more than once (#31)
+* Pass all outputs from a step, instead of just one (#27)
+* Improve listing of workflow steps
+
 ## 0.3.1
 
 ### Added